Move Category Enum Class to Enums folder
