+++ src/main/java/classes/Foods/Foods.java
@@ -2,7 +2,7 @@ package classes.Foods;
 
 import java.util.HashMap;
 import java.util.LinkedList;
-import classes.*;
+import classes.Enums.Category;
 
 
 public class Foods
