+++ src/main/java/classes/Users/Users.java
@@ -49,3 +49,14 @@ public class Users {
         }
     }
 }
+
+
+
+// TASKS -->
+// CHANGE COMMENTS
+// CREATE A TABLE FORMAT FOR CUSTOMER DETAILS
+// IMPLEMENT LOGIN
+// IMPLEMENT REGISTER
+// PERFORM THE INHERITANCE OPERATION IN LOGIN AND REGISTER
+// CHECK THE FILE I/O OPERATION TO STORE USERS' DETAILS AFTER REGISTRATION PROCESS
+// CHECK THE FILE I/O OPERATION TO READ USERS' DETAILS FOR LOGIN PROCESS
